+++ rocketpy/rocket/aero_surface.py
@@ -629,8 +629,8 @@ class Fins(AeroSurface):
             Otherwise, if tuple, fins will be considered as airfoils. The
             tuple's first item specifies the airfoil's lift coefficient
             by angle of attack and must be either a .csv, .txt, ndarray
-            or callable. The .csv and .txt files must contain no headers
-            and the first column must specify the angle of attack, while
+            or callable. The .csv and .txt files can contain a single line
+            header and the first column must specify the angle of attack, while
             the second column must specify the lift coefficient. The
             ndarray should be as [(x0, y0), (x1, y1), (x2, y2), ...]
             where x0 is the angle of attack and y0 is the lift coefficient.
@@ -1037,8 +1037,8 @@ class TrapezoidalFins(Fins):
             Otherwise, if tuple, fins will be considered as airfoils. The
             tuple's first item specifies the airfoil's lift coefficient
             by angle of attack and must be either a .csv, .txt, ndarray
-            or callable. The .csv and .txt files must contain no headers
-            and the first column must specify the angle of attack, while
+            or callable. The .csv and .txt files can contain a single line
+            header and the first column must specify the angle of attack, while
             the second column must specify the lift coefficient. The
             ndarray should be as [(x0, y0), (x1, y1), (x2, y2), ...]
             where x0 is the angle of attack and y0 is the lift coefficient.
@@ -1379,8 +1379,8 @@ class EllipticalFins(Fins):
             Otherwise, if tuple, fins will be considered as airfoils. The
             tuple's first item specifies the airfoil's lift coefficient
             by angle of attack and must be either a .csv, .txt, ndarray
-            or callable. The .csv and .txt files must contain no headers
-            and the first column must specify the angle of attack, while
+            or callable. The .csv and .txt files can contain a single line
+            header and the first column must specify the angle of attack, while
             the second column must specify the lift coefficient. The
             ndarray should be as [(x0, y0), (x1, y1), (x2, y2), ...]
             where x0 is the angle of attack and y0 is the lift coefficient.